Refactor: added a new package org.jline.console
