+++ builtins/src/main/java/org/jline/builtins/Builtins.java
@@ -22,15 +22,16 @@ import java.util.stream.Collectors;
 import org.jline.builtins.Completers.FilesCompleter;
 import org.jline.builtins.Completers.OptDesc;
 import org.jline.builtins.Completers.OptionCompleter;
-import org.jline.builtins.Completers.SystemCompleter;
 import org.jline.builtins.Options.HelpException;
-import org.jline.builtins.Widgets.ArgDesc;
-import org.jline.builtins.Widgets.CmdDesc;
+import org.jline.console.ArgDesc;
+import org.jline.console.CmdDesc;
+import org.jline.console.ConfigurationPath;
 import org.jline.reader.*;
 import org.jline.reader.LineReader.Option;
 import org.jline.reader.impl.completer.ArgumentCompleter;
 import org.jline.reader.impl.completer.NullCompleter;
 import org.jline.reader.impl.completer.StringsCompleter;
+import org.jline.reader.impl.completer.SystemCompleter;
 import org.jline.terminal.Terminal;
 import org.jline.utils.AttributedString;
