+++ builtins/src/main/java/org/jline/builtins/Commands.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2019, the original author or authors.
+ * Copyright (c) 2002-2020, the original author or authors.
  *
  * This software is distributable under the BSD license. See the terms of the
  * BSD license in the documentation provided with this software.
@@ -44,9 +44,9 @@ import org.jline.builtins.Options;
 import org.jline.builtins.Options.HelpException;
 import org.jline.builtins.Source.StdInSource;
 import org.jline.builtins.Source.URLSource;
+import org.jline.console.ConfigurationPath;
 import org.jline.keymap.KeyMap;
 import org.jline.reader.Binding;
-import org.jline.reader.ConfigurationPath;
 import org.jline.reader.Highlighter;
 import org.jline.reader.History;
 import org.jline.reader.LineReader;
