+++ groovy/src/main/java/org/jline/script/GroovyEngine.java
@@ -14,8 +14,8 @@ import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.jline.console.ScriptEngine;
 import org.jline.groovy.Utils;
-import org.jline.reader.ScriptEngine;
 
 import groovy.lang.Binding;
 import groovy.lang.Closure;
